Fix OSX build because of missing GCC variable. [skip ci]

+++ recipe/build.sh
@@ -4,7 +4,9 @@
 CXXFLAGS="${CXXFLAGS} -D_GLIBCXX_USE_CXX11_ABI=0"
 
 # fix problem building with conda-forge python that doesn't use Anaconda compilers
-ln -s "$GCC" "$PREFIX/bin/gcc"
+if [ ! -z "$GCC" ]; then
+    ln -s "$GCC" "$PREFIX/bin/gcc"
+fi
 
 mkdir build
 cd build
@@ -19,4 +21,7 @@ rm -r /tmp/hdf5
 make install
 
 # clean up python build fix
-rm "$PREFIX/bin/gcc"
+if [ ! -z "$GCC" ]; then
+    rm "$PREFIX/bin/gcc"
+fi
+
